# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Salin API",
    "description": "Upload and transcript orchestration for Salin.",
    "version": "0.1.0"
  },
  "paths": {
    "/recordings": {
      "get": {
        "summary": "List Recordings",
        "operationId": "list_recordings_recordings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Recording",
        "operationId": "create_recording_recordings_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_recording_recordings_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}": {
      "get": {
        "summary": "Get Recording",
        "operationId": "get_recording_recordings__recording_id__get",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/retry": {
      "post": {
        "summary": "Retry Recording",
        "operationId": "retry_recording_recordings__recording_id__retry_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/notes/generate": {
      "post": {
        "summary": "Generate Notes",
        "operationId": "generate_notes_recordings__recording_id__notes_generate_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/notes": {
      "put": {
        "summary": "Update Notes",
        "operationId": "update_notes_recordings__recording_id__notes_put",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/speakers/rename": {
      "put": {
        "summary": "Rename Speaker",
        "operationId": "rename_speaker_recordings__recording_id__speakers_rename_put",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptSegmentsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/transcript-segments/{segment_id}/speaker": {
      "put": {
        "summary": "Update Segment Speaker",
        "operationId": "update_segment_speaker_recordings__recording_id__transcript_segments__segment_id__speaker_put",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Segment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentSpeakerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptSegmentsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactUrls": {
        "properties": {
          "original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original"
          },
          "normalized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized"
          }
        },
        "type": "object",
        "title": "ArtifactUrls"
      },
      "Body_create_recording_recordings_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageOption"
          },
          "processing_mode": {
            "$ref": "#/components/schemas/ProcessingMode"
          },
          "speaker_count": {
            "$ref": "#/components/schemas/SpeakerCount"
          }
        },
        "type": "object",
        "required": [
          "file",
          "language",
          "processing_mode",
          "speaker_count"
        ],
        "title": "Body_create_recording_recordings_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LanguageOption": {
        "type": "string",
        "enum": [
          "auto",
          "tl",
          "en"
        ],
        "title": "LanguageOption"
      },
      "NotesGenerationResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "notes": {
            "$ref": "#/components/schemas/GeneratedNotesSummary"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "notes"
        ],
        "title": "NotesGenerationResponse"
      },
...
      "NotesUpdateRequest": {
        "properties": {
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points"
          },
          "decisions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Decisions"
          },
          "action_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Items"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "key_points",
          "decisions",
          "action_items",
          "questions"
        ],
        "title": "NotesUpdateRequest"
      },
      "ProcessingJobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "last_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Provider"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "stage",
          "retryable",
          "retry_count",
          "error_message",
          "last_provider",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at"
        ],
        "title": "ProcessingJobSummary"
      },
      "ProcessingMode": {
        "type": "string",
        "enum": [
          "fast",
          "accurate"
        ],
        "title": "ProcessingMode"
      },
      "RecordingCreateResponse": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingSummary"
          },
          "job": {
            "$ref": "#/components/schemas/ProcessingJobSummary"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "job"
        ],
        "title": "RecordingCreateResponse"
      },
      "RecordingDetailResponse": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingSummary"
          },
          "job": {
            "$ref": "#/components/schemas/ProcessingJobSummary"
          },
          "transcript_segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentSummary"
            },
            "type": "array",
            "title": "Transcript Segments"
          },
          "artifact_urls": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactUrls"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "$ref": "#/components/schemas/GeneratedNotesSummary"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "job",
          "transcript_segments",
          "notes"
        ],
        "title": "RecordingDetailResponse"
      },
      "RecordingListItemSummary": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingSummary"
          },
          "job": {
            "$ref": "#/components/schemas/ProcessingJobSummary"
          },
          "notes": {
            "$ref": "#/components/schemas/GeneratedNotesSummary"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "job",
          "notes"
        ],
        "title": "RecordingListItemSummary"
      },
      "RecordingListResponse": {
        "properties": {
          "recordings": {
            "items": {
              "$ref": "#/components/schemas/RecordingListItemSummary"
            },
            "type": "array",
            "title": "Recordings"
          }
        },
        "type": "object",
        "required": [
          "recordings"
        ],
        "title": "RecordingListResponse"
      },
      "RecordingSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageOption"
          },
          "processing_mode": {
            "$ref": "#/components/schemas/ProcessingMode"
          },
          "speaker_count": {
            "$ref": "#/components/schemas/SpeakerCount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "content_type",
          "file_size",
          "language",
          "processing_mode",
          "speaker_count",
          "created_at",
          "updated_at"
        ],
        "title": "RecordingSummary"
      },
      "RetryResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "job": {
            "$ref": "#/components/schemas/ProcessingJobSummary"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "job"
        ],
        "title": "RetryResponse"
      },
      "SegmentSpeakerUpdateRequest": {
        "properties": {
          "speaker_label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Speaker Label"
          }
        },
        "type": "object",
        "required": [
          "speaker_label"
        ],
        "title": "SegmentSpeakerUpdateRequest"
      },
      "SpeakerCount": {
        "type": "string",
        "enum": [
          "auto",
          "1",
          "2",
          "3",
          "4",
          "5_plus"
        ],
        "title": "SpeakerCount"
      },
      "SpeakerRenameRequest": {
        "properties": {
          "from_label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "From Label"
          },
          "to_label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "To Label"
          }
        },
        "type": "object",
        "required": [
          "from_label",
          "to_label"
        ],
        "title": "SpeakerRenameRequest"
      },
      "TranscriptSegmentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "speaker_label": {
            "type": "string",
            "title": "Speaker Label"
          },
          "speaker_estimated": {
            "type": "boolean",
            "title": "Speaker Estimated"
          },
          "source_provider": {
            "type": "string",
            "title": "Source Provider"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "index",
          "start_ms",
          "end_ms",
          "text",
          "speaker_label",
          "speaker_estimated",
          "source_provider"
        ],
        "title": "TranscriptSegmentSummary"
      },
      "TranscriptSegmentsUpdateResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "transcript_segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentSummary"
            },
            "type": "array",
            "title": "Transcript Segments"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "transcript_segments"
        ],
        "title": "TranscriptSegmentsUpdateResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}